CHANGED: database.py - filter variables are now filled via data from the XML reader object NEW: filters.xml - xml file containing the data for the various filters NEW: xmlfilters.py - simple xml reader that translates filters.xml into usable python lists
Note: A malformed XML file will cause the program to crash. Due to the existing nature of how the filters are loaded into code, recoding of explicit list calls in card_filters.py would have to be done.

Please test to ensure filter functionality is the same in the client.